chore: Rename BisUtils.Bank to BisUtils.RVBank
