+++ BisUtils.sln
@@ -16,7 +16,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lib", "lib", "{3BC325D5-711
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BisUtils.Core", "src\BisUtils.Core\BisUtils.Core.csproj", "{F75A0623-0022-4204-9FDE-BFC39583C413}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BisUtils.Bank", "src\File Formats\BisUtils.Bank\BisUtils.Bank.csproj", "{F53CAC57-01A3-4A71-88D6-19146D271933}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BisUtils.RVBank", "src\File Formats\BisUtils.RVBank\BisUtils.RVBank.csproj", "{F53CAC57-01A3-4A71-88D6-19146D271933}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BisUtils.Param", "src\File Formats\Languages\BisUtils.Param\BisUtils.Param.csproj", "{1C3D2806-0E33-4816-BF09-3A84CF50AA80}"
 EndProject
