+++ src/File Formats/BisUtils.RVBank/Alerts/Errors/PboEmptyPropertyNameError.cs
@@ -1,4 +1,4 @@
-﻿namespace BisUtils.Bank.Alerts.Errors;
+﻿namespace BisUtils.RVBank.Alerts.Errors;
 
 using FResults.Reasoning;
 using Model;
