+++ src/File Formats/BisUtils.RVBank/Alerts/Warnings/PboImproperMimeWarning.cs
@@ -1,11 +1,11 @@
-﻿namespace BisUtils.Bank.Alerts.Warnings;
+﻿namespace BisUtils.RVBank.Alerts.Warnings;
 
 using FResults.Reasoning;
 using Model.Stubs;
 
-public class PboUnnamedEntryWarning : WarningBase
+public class PboImproperMimeWarning : WarningBase
 {
-    public PboUnnamedEntryWarning(bool isError = true, Type? type = null)
+    public PboImproperMimeWarning(bool isError = true, Type? type = null)
     {
         IsError = isError;
         AlertScope = type ?? typeof(PboEntry);
@@ -13,7 +13,7 @@ public class PboUnnamedEntryWarning : WarningBase
 
     public override string? AlertName
     {
-        get => "UnnamedEntry";
+        get => "ImproperMime";
         init => throw new NotSupportedException();
     }
 
@@ -31,7 +31,7 @@ public class PboUnnamedEntryWarning : WarningBase
 
     public override string? Message
     {
-        get => "A pbo entry was found with no name.";
+        get => "An improper mime was found on an entry";
         set => throw new NotSupportedException();
     }
 }
